[nrf fromtree] Bluetooth: hci: Fix wrong CTE request interval type
Request interval is a number of connection events that
is used to periodically run CTE request control procedure.
BT 5.3 Core Specification defines it as 2 octets long.
It had wrong type uint8_t. Changed to correct one uint16_t.

The commit also changes type of cte_rsp_en field of lll_df_conn_tx_cfg
to state that it is a boolean flag.

(cherry picked from commit 2ac5f7574b87885f3baf432b1559698000ec8517)

+++ include/bluetooth/hci.h
@@ -1607,7 +1607,7 @@ struct bt_hci_rp_le_set_conn_cte_tx_params {
 struct bt_hci_cp_le_conn_cte_req_enable {
 	uint16_t handle;
 	uint8_t  enable;
-	uint8_t  cte_request_interval;
+	uint16_t cte_request_interval;
 	uint8_t  requested_cte_length;
 	uint8_t  requested_cte_type;
 } __packed;
